Add files via upload
This modified project was based on the one presented on the presentation. Add the function of recording the whole week even a year.
Realize way: keep the txt file in src file. Therefore, after user typing in his/his name, the project can make a judgment of whether this is an old user. If it is, use the kept txt file, else, create a new txt file